+++ R/req-headers.R
@@ -9,6 +9,10 @@
 #'   * Use `NULL` to reset a value to httr's default
 #'   * Use `""` to remove a header
 #'   * Use a character vector to repeat a header.
+#' @param .redact If `TRUE`, all the added headers are redacted. Can also be a
+#'   named list containing `TRUE` or `FALSE` declaring whether or not to redact
+#'   a particular header. If a named list is provided, the default for any
+#'   unspecified header is `TRUE`.
 #' @returns A modified HTTP [request].
 #' @export
 #' @examples
@@ -47,9 +51,44 @@
 #'    req_headers(!!!headers, HeaderThree = "three") %>%
 #'    req_dry_run()
 #'
-req_headers <- function(.req, ...) {
+req_headers <- function(.req, ..., .redact = FALSE) {
   check_request(.req)
 
-  .req$headers <- modify_list(.req$headers, ...)
+  headers <- list2(...)
+  header_names <- names2(headers)
+  .redact <- unlist(check_list_of_bool(.redact, header_names))
+  to_redact <- names(.redact)[.redact]
+  to_unredact <- names(.redact)[!.redact]
+
+  redact_out <- attr(.req$headers, "redact") %||% to_redact
+  redact_out <- setdiff(redact_out, to_unredact)
+  redact_out <- union(redact_out, to_redact)
+  .req$headers <- modify_list(.req$headers, !!!headers)
+
+  attr(.req$headers, "redact") <- redact_out
+
   .req
+}
+
+check_list_of_bool <- function(x, names, arg = caller_arg(x), call = caller_env()) {
+  if (is_bool(x)) {
+    rep_named(names, x)
+  } else if (is_bare_list(x)) {
+    check_unique_names(x, arg = arg, call = call)
+    x[intersect(names(x), names)]
+  } else  {
+    cli::cli_abort(
+      "{.arg {arg}} must be a list or a single `TRUE` or `FALSE`.",
+      call = call
+    )
+  }
+}
+
+check_unique_names <- function(x, arg = caller_arg(x), call = caller_env()) {
+  if (length(x) > 0L && !is_named(x)) {
+    cli::cli_abort("All elements of {.arg {arg}} must be named.", call = call)
+  }
+  if (anyDuplicated(names(x))) {
+    cli::cli_abort("The names of {.arg {arg}} must be unique.", call = call)
+  }
 }